CanvasUtil removed
All CanvasUtil functions ported into CanvasWrapper using PyCanvas instead of direct POST/GET calls. Some testing performed, but needs to be tested more extensively
